racsAWSPipelineForARTLab/python/src/requirements.txt
audeer==1.19.0
audformat==0.16.0
audinterface==0.10.2
audiofile==1.2.1
audmath==1.2.1
audobject==0.7.9
audresample==1.2.1
backports.tempfile==1.0
backports.weakref==1.0.post1
blis==0.7.7
brotlipy==0.7.0
catalogue==2.0.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.0.4
cmake==3.25.2
colorama==0.4.6
cryptography==39.0.1
cymem==2.0.6
ffmpeg-python==0.2.0
filelock==3.9.0
flit_core==3.6.0
future==0.18.3
huggingface-hub==0.13.1
idna==3.4
importlib-metadata==6.0.0
iso-639==0.4.5
iso3166==2.1.1
Jinja2==3.1.2
joblib==1.2.0
langcodes==3.3.0
lit==15.0.7
llvmlite==0.39.1
MarkupSafe==2.1.1
matplotlib
more-itertools==9.1.0
murmurhash==1.0.7
nltk==3.8.1
numba==0.56.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
openai-whisper==20230308
opensmile==2.4.2
oyaml==1.0
packaging==22.0
pandas==1.5.3
pathy==0.10.1
praat-parselmouth==0.4.3
preshed==3.0.6
pycparser==2.21
pydantic==1.8.2
pyOpenSSL==23.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
shellingham==1.5.0
six==1.16.0
smart-open==5.2.1
soundfile==0.12.1
spacy==3.3.1
spacy-legacy==3.0.9
spacy-loggers==1.0.1
SPLAT-library
srsly==2.4.6
thinc==8.0.15
tokenizers==0.13.2
torch==1.13.1
torchaudio==0.13.1
tqdm==4.64.1
transformers==4.26.1
triton==2.0.0
typer==0.4.1
typing_extensions==4.4.0
urllib3==1.26.14
wasabi==0.9.1
zipp==3.15.0
